+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add support for measurement ingestion. ([#724](https://github.com/getsentry/relay/pull/724), [#785](https://github.com/getsentry/relay/pull/785))
 - Add support for scrubbing UTF-16 data in attachments ([#742](https://github.com/getsentry/relay/pull/742), [#784](https://github.com/getsentry/relay/pull/784), [#787](https://github.com/getsentry/relay/pull/787))
+- Add upstream request metric. ([#793](https://github.com/getsentry/relay/pull/793))
 
 **Bug Fixes**: